Split platform module more
