+++ maplibre-apple/src/lib.rs
@@ -1,0 +1,16 @@
+use maplibre::{MapBuilder, ScheduleMethod, TokioScheduleMethod};
+use maplibre::window::FromWindow;
+pub use std::time::Instant;
+
+#[cfg(not(any(target_os = "macos", target_os = "ios")))]
+compile_error!("maplibre-apple works only on macOS and iOS.");
+
+#[no_mangle]
+pub fn maplibre_apple_main() {
+    env_logger::init_from_env(env_logger::Env::default().default_filter_or("info"));
+
+    MapBuilder::from_window("A fantastic window!")
+        .with_schedule_method(ScheduleMethod::Tokio(TokioScheduleMethod::new()))
+        .build()
+        .run_sync();
+}
